back/info/requirements.txt
easyocr>=1.7.0
pika>=1.3.2
minio>=7.0.0
requests>=2.25.0
python-dotenv>=0.19.0
opencv-python>=4.5.0
numpy>=1.20.0

# Для работы с JSON и типами
python-dateutil>=2.8.2
typing-extensions>=4.0.0

# Для логирования
logging
